deployment_resume/requirements.txt
streamlit
scikit-learn
nltk
numpy
panda
pdf2docx
PyPDF2